+++ service/Models/WeightInputCommandModel.cs
@@ -6,6 +6,6 @@ public class WeightInputCommandModel
 {
     [Range(20, 500)] public required decimal Weight { get; set; }
     [Required] public required DateTime Date { get; set; }
-    public float? BodyFatPercentage { get; set; }
-    public float? SkeletalMuscleWeight { get; set; }
+    public decimal? BodyFatPercentage { get; set; }
+    public decimal? SkeletalMuscleWeight { get; set; }
 }
